requirements.txt
base58==2.1.1
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
six==1.16.0
solana==0.29.0
solders==0.14.2
